Changed KqueueManager::LocalToSystem and SystemToLocal to use uint16_t to better align with OSX's API. Add some extra logging and error handling in KqueueManager. Fixed bug with handling of timeout in KqueueManager::Wait(). KqueueManager::AddSocket() fixed incorrect usage of EV_SET. Tweaked KqueueManager::RemoveSocket(). Better logging and error handling in Socket::SetReuseAddr().

+++ externals/Library/Event/include/KqueueManager.hxx
@@ -24,8 +24,8 @@ class KqueueManager : public Manager {
 	private:
 		int kfd;
 		struct kevent event_list[max_events];
-		uint32_t LocalToSystem(uint32_t filter_in);
-		uint32_t SystemToLocal(uint32_t filter_in);
+		int16_t LocalToSystem(int16_t filter_in);
+		int16_t SystemToLocal(int16_t filter_in);
 
 
 	public:
